// src/types/IBook.ts
export interface IBook {
  _created: number;
  _data_type: string;
  _is_deleted: boolean;
  _modified: number;
  _self_link: string;
  _user: string;
  _uuid: string;
  author: string;
  availability: number;
  cover: string;
  description: string;
  language: string;
  pages: number;
  price: number;
  publisher: string;
  title: string;
}